+++ app/controllers/cookies_controller.rb
@@ -2,18 +2,13 @@ class CookiesController < ApplicationController
   skip_before_action :redirect_unauthenticated_user
 
   def edit
-    @cookie_form = CookiesForm.new(accept_optional_cookies: get_cookie_value)
   end
 
   def update
-    cookie_value = cookies_params[:accept_optional_cookies]
-    cookies[:ACCEPT_OPTIONAL_COOKIES] = {value: cookie_value, expires: 1.year}
-  end
+    accept_optional_cookies = cookies_params[:accept_optional_cookies]
+    cookies[:ACCEPT_OPTIONAL_COOKIES] = {value: accept_optional_cookies, expires: 1.year} unless accept_optional_cookies.nil?
 
-  private def get_cookie_value
-    return false unless cookies[:ACCEPT_OPTIONAL_COOKIES].present?
-
-    JSON.parse(cookies[:ACCEPT_OPTIONAL_COOKIES])
+    redirect_back_or_to(cookies_path, notice: "Your cookie preferences have been updated")
   end
 
   private def cookies_params